+++ auto_process_ngs/test/tenx/test_utils.py
@@ -1,27 +1,13 @@
 #######################################################################
-# Tests for tenx_genomics_utils.py module
+# Tests for tenx_genomics/utils.py module
 #######################################################################
 
 import unittest
 import tempfile
 import os
 import shutil
-from bcftbx.mock import MockIlluminaRun
 from bcftbx.mock import RunInfoXml
-from auto_process_ngs.analysis import AnalysisProject
-from auto_process_ngs.mock import MockBcl2fastq2Exe
-from auto_process_ngs.mock import MockCellrangerExe
-from auto_process_ngs.mock import MockAnalysisDir
-from auto_process_ngs.mock import MockAnalysisProject
-from auto_process_ngs.mock10xdata import METRICS_SUMMARY
-from auto_process_ngs.mock10xdata import ATAC_SUMMARY
-from auto_process_ngs.mock10xdata import ATAC_SUMMARY_2_0_0
-from auto_process_ngs.mock10xdata import CELLPLEX_METRICS_SUMMARY
-from auto_process_ngs.mock10xdata import MULTIOME_SUMMARY
-from auto_process_ngs.mock10xdata import MULTIOME_SUMMARY_2_0_0
-from auto_process_ngs.mock10xdata import MULTIOME_LIBRARIES
-from auto_process_ngs.mock10xdata import MULTIOME_LIBRARIES_NO_RUN
-from auto_process_ngs.tenx_genomics_utils import *
+from auto_process_ngs.tenx.utils import *
 
 # Set to False to keep test output dirs
 REMOVE_TEST_OUTPUTS = True
@@ -567,369 +553,3 @@ class TestSpacerangerInfo(unittest.TestCase):
         spaceranger = self._make_mock_spaceranger_131()
         self.assertEqual(spaceranger_info(name='spaceranger'),
                          (spaceranger,'spaceranger','1.3.1'))
-
-
-class TestGexSummary(unittest.TestCase):
-    """
-    Tests for the 'GexSummary' class
-    """
-    def setUp(self):
-        # Create a temp working dir
-        self.wd = tempfile.mkdtemp(suffix='TestGexSummary')
-
-    def tearDown(self):
-        # Remove the temporary test directory
-        if REMOVE_TEST_OUTPUTS:
-            shutil.rmtree(self.wd)
-            
-    def test_gex_summary(self):
-        """GexSummary: check metrics are extracted from CSV file
-        """
-        summary_csv = os.path.join(self.wd,"metrics_summary.csv")
-        with open(summary_csv,'w') as fp:
-            fp.write(METRICS_SUMMARY)
-        m = GexSummary(summary_csv)
-        self.assertEqual(m.estimated_number_of_cells,2272)
-        self.assertEqual(m.mean_reads_per_cell,107875)
-        self.assertEqual(m.median_genes_per_cell,1282)
-
-class TestAtacSummary(unittest.TestCase):
-    """
-    Tests for the 'AtacSummary' class
-    """
-    def setUp(self):
-        # Create a temp working dir
-        self.wd = tempfile.mkdtemp(suffix='TestAtacSummary')
-
-    def tearDown(self):
-        # Remove the temporary test directory
-        if REMOVE_TEST_OUTPUTS:
-            shutil.rmtree(self.wd)
-
-    def test_atac_summary_pre_2_0_0(self):
-        """AtacSummary: get metrics from summary.csv (Cellranger ATAC pre-2.0.0)
-        """
-        summary_csv = os.path.join(self.wd,"summary.csv")
-        with open(summary_csv,'w') as fp:
-            fp.write(ATAC_SUMMARY)
-        s = AtacSummary(summary_csv)
-        self.assertEqual(s.version,"1.0.1")
-        self.assertEqual(s.cells_detected,6748)
-        self.assertEqual(s.annotated_cells,5682)
-        self.assertEqual(s.median_fragments_per_cell,16119.5)
-        self.assertEqual(s.frac_fragments_overlapping_targets,
-                         0.575082094792)
-        self.assertEqual(s.frac_fragments_overlapping_peaks,
-                         0.556428090013)
-        self.assertEqual(s.tss_enrichment_score,6.91438390781)
-        self.assertRaises(AttributeError,
-                          getattr,
-                          s,'estimated_number_of_cells')
-
-    def test_atac_summary_2_0_0(self):
-        """AtacSummary: get metrics from summary.csv (Cellranger ATAC 2.0.0)
-        """
-        summary_csv = os.path.join(self.wd,"summary.csv")
-        with open(summary_csv,'w') as fp:
-            fp.write(ATAC_SUMMARY_2_0_0)
-        s = AtacSummary(summary_csv)
-        self.assertEqual(s.version,"2.0.0")
-        self.assertEqual(s.estimated_number_of_cells,3582)
-        self.assertEqual(s.median_fragments_per_cell,51354.5)
-        self.assertEqual(s.frac_fragments_overlapping_peaks,0.4856)
-        self.assertEqual(s.tss_enrichment_score,6.8333)
-        self.assertRaises(AttributeError,
-                          getattr,
-                          s,'cells_detected')
-        self.assertRaises(AttributeError,
-                          getattr,
-                          s,'annotated_cells')
-        self.assertRaises(AttributeError,
-                          getattr,
-                          s,'frac_fragments_overlapping_targets')
-
-class TestMultiomeSummary(unittest.TestCase):
-    """
-    Tests for the 'MultiomeSummary' class
-    """
-    def setUp(self):
-        # Create a temp working dir
-        self.wd = tempfile.mkdtemp(suffix='TestMultiomeSummary')
-
-    def tearDown(self):
-        # Remove the temporary test directory
-        if REMOVE_TEST_OUTPUTS:
-            shutil.rmtree(self.wd)
-
-    def test_atac_summary_multiome_arc_1_0_0(self):
-        """MultiomeSummary: check metrics are extracted from CSV file (Cellranger ARC 1.0.0)
-        """
-        summary_csv = os.path.join(self.wd,"summary.csv")
-        with open(summary_csv,'wt') as fp:
-            fp.write(MULTIOME_SUMMARY)
-        s = MultiomeSummary(summary_csv)
-        self.assertEqual(s.estimated_number_of_cells,744)
-        self.assertEqual(
-            s.atac_median_high_quality_fragments_per_cell,8079)
-        self.assertEqual(s.gex_median_genes_per_cell,1490)
-
-    def test_atac_summary_multiome_arc_2_0_0(self):
-        """MultiomeSummary: check metrics are extracted from CSV file (Cellranger ARC 2.0.0)
-        """
-        summary_csv = os.path.join(self.wd,"summary.csv")
-        with open(summary_csv,'wt') as fp:
-            fp.write(MULTIOME_SUMMARY_2_0_0)
-        s = MultiomeSummary(summary_csv)
-        self.assertEqual(s.estimated_number_of_cells,785)
-        self.assertEqual(
-            s.atac_median_high_quality_fragments_per_cell,9.0)
-        self.assertEqual(s.gex_median_genes_per_cell,15.0)
-
-class TestMultiplexSummary(unittest.TestCase):
-    """
-    Tests for the 'MultiplexSummary' class
-    """
-    def setUp(self):
-        # Create a temp working dir
-        self.wd = tempfile.mkdtemp(suffix='TestMultiplexSummary')
-
-    def tearDown(self):
-        # Remove the temporary test directory
-        if REMOVE_TEST_OUTPUTS:
-            shutil.rmtree(self.wd)
-
-    def test_multiplex_summary(self):
-        """MultiplexSummary: check metrics are extracted from CSV file
-        """
-        summary_csv = os.path.join(self.wd,"metrics_summary.csv")
-        with open(summary_csv,'w') as fp:
-            fp.write(CELLPLEX_METRICS_SUMMARY)
-        m = MultiplexSummary(summary_csv)
-        self.assertEqual(m.cells,5175)
-        self.assertEqual(m.median_reads_per_cell,20052)
-        self.assertEqual(m.median_genes_per_cell,3086)
-        self.assertEqual(m.total_genes_detected,21260)
-        self.assertEqual(m.median_umi_counts_per_cell,10515)
-
-class TestMultiomeLibraries(unittest.TestCase):
-    """
-    Tests for the 'MultiomeLibraries' class
-    """
-    def setUp(self):
-        # Create a temp working dir
-        self.wd = tempfile.mkdtemp(suffix='TestMultiomeLibraries')
-
-    def tearDown(self):
-        # Remove the temporary test directory
-        if REMOVE_TEST_OUTPUTS:
-            shutil.rmtree(self.wd)
-
-    def test_multiome_libraries(self):
-        """MultiomeLibraries: read in data from a file
-        """
-        multiome_libraries_info = os.path.join(self.wd,
-                                               "10x_multiome_libraries.info")
-        with open(multiome_libraries_info,'wt') as fp:
-            fp.write(MULTIOME_LIBRARIES)
-        m = MultiomeLibraries(multiome_libraries_info)
-        # Check list of samples
-        self.assertEqual(m.local_samples,['PB1_ATAC','PB2_ATAC'])
-        # Check lists of linked samples
-        self.assertEqual(m.linked_samples('PB1_ATAC'),
-                         ['NEXTSEQ_210111/12:PB_GEX/PB1_GEX',])
-        self.assertEqual(m.linked_samples('PB2_ATAC'),
-                         ['NEXTSEQ_210111/12:PB_GEX/PB2_GEX',])
-
-    def test_multiome_libraries_linked_projects(self):
-        """MultiomeLibraries: return linked projects
-        """
-        multiome_libraries_info = os.path.join(self.wd,
-                                               "10x_multiome_libraries.info")
-        with open(multiome_libraries_info,'wt') as fp:
-            fp.write(MULTIOME_LIBRARIES)
-        m = MultiomeLibraries(multiome_libraries_info)
-        # Create run with linked project
-        run = MockAnalysisDir("210111_NB01234_00012_ABXXXXX",
-                              "nextseq",
-                              project_metadata={
-                                  'PB_GEX': {
-                                      'Library type': 'GEX',
-                                  },
-                              },
-                              lanes=(1,2,3,4),
-                              top_dir=self.wd)
-        run.add_fastq_batch('PB_GEX','PB1_GEX','PB1_GEX_S1',lanes=(1,2,3,4))
-        run.add_fastq_batch('PB_GEX','PB2_GEX','PB2_GEX_S2',lanes=(1,2,3,4))
-        run.create()
-        # Check list of linked projects
-        linked_projects = m.linked_projects()
-        self.assertEqual(len(linked_projects),1)
-        self.assertEqual(linked_projects[0].dirn,
-                         os.path.join(self.wd,
-                                      "210111_NB01234_00012_ABXXXXX_analysis",
-                                      "PB_GEX"))
-
-    def test_multiome_libraries_linked_projects_same_run(self):
-        """MultiomeLibraries: return linked projects from same run
-        """
-        # Create run with linked projects
-        run = MockAnalysisDir("210111_NB01234_00012_ABXXXXX",
-                              "nextseq",
-                              project_metadata={
-                                  'PB_ATAC': {
-                                      'Library type': 'ATAC',
-                                  },
-                                  'PB_GEX': {
-                                      'Library type': 'GEX',
-                                  },
-                              },
-                              lanes=(1,2,3,4),
-                              top_dir=self.wd)
-        run.add_fastq_batch('PB_ATAC','PB1_ATAC','PB1_ATAC_S1',lanes=(1,2,3,4))
-        run.add_fastq_batch('PB_ATAC','PB2_ATAC','PB2_ATAC_S2',lanes=(1,2,3,4))
-        run.add_fastq_batch('PB_GEX','PB1_GEX','PB1_GEX_S3',lanes=(1,2,3,4))
-        run.add_fastq_batch('PB_GEX','PB2_GEX','PB2_GEX_S4',lanes=(1,2,3,4))
-        run.create()
-        # Put multiome libraries file into ATAC project
-        multiome_libraries_info = os.path.join(run.dirn,
-                                               "PB_ATAC",
-                                               "10x_multiome_libraries.info")
-        with open(multiome_libraries_info,'wt') as fp:
-            fp.write(MULTIOME_LIBRARIES_NO_RUN)
-        m = MultiomeLibraries(multiome_libraries_info)
-        # Check list of linked projects
-        linked_projects = m.linked_projects()
-        self.assertEqual(len(linked_projects),1)
-        self.assertEqual(linked_projects[0].dirn,
-                         os.path.join(self.wd,
-                                      "210111_NB01234_00012_ABXXXXX_analysis",
-                                      "PB_GEX"))
-
-    def test_multiome_libraries_write_libraries_csv(self):
-        """MultiomeLibraries: write 'libraries.csv' files
-        """
-        multiome_libraries_info = os.path.join(self.wd,
-                                               "10x_multiome_libraries.info")
-        with open(multiome_libraries_info,'wt') as fp:
-            fp.write(MULTIOME_LIBRARIES)
-        m = MultiomeLibraries(multiome_libraries_info)
-        # Create run with linked project
-        run = MockAnalysisDir("210111_NB01234_00012_ABXXXXX",
-                              "nextseq",
-                              project_metadata={
-                                  'PB_GEX': {
-                                      'Library type': 'GEX',
-                                  },
-                              },
-                              lanes=(1,2,3,4),
-                              top_dir=self.wd)
-        run.add_fastq_batch('PB_GEX','PB1_GEX','PB1_GEX_S1',lanes=(1,2,3,4))
-        run.add_fastq_batch('PB_GEX','PB2_GEX','PB2_GEX_S1',lanes=(1,2,3,4))
-        run.create()
-        # Write libraries.csv file
-        libraries_csv = os.path.join(self.wd,"libraries.csv")
-        self.assertFalse(os.path.exists(libraries_csv))
-        m.write_libraries_csv('PB1_ATAC',
-                              '/data/PB_ATAC/fastqs',
-                              'ATAC',
-                              filen=libraries_csv)
-        self.assertTrue(os.path.exists(libraries_csv))
-        with open(libraries_csv,'rt') as fp:
-            contents = fp.read()
-            self.assertEqual(contents,
-                             """fastqs,sample,library_type
-/data/PB_ATAC/fastqs,PB1_ATAC,Chromatin Accessibility
-%s/210111_NB01234_00012_ABXXXXX_analysis/PB_GEX/fastqs,PB1_GEX,Gene Expression
-""" % self.wd)
-
-class TestCellrangerMultiConfigCsv(unittest.TestCase):
-    """
-    Tests for the 'CellrangerMultiConfigCsv' class
-    """
-    def setUp(self):
-        # Create a temp working dir
-        self.wd = tempfile.mkdtemp(suffix='TestCellrangerMultiConfigCsv')
-    def tearDown(self):
-        # Remove the temporary test directory
-        if REMOVE_TEST_OUTPUTS:
-            shutil.rmtree(self.wd)
-
-    def test_cellranger_multi_config_csv(self):
-        """
-        CellrangerMultiConfigCsv: check data is extracted from config.csv
-        """
-        with open(os.path.join(self.wd,"10x_multi_config.csv"),'wt') as fp:
-            fp.write("""[gene-expression]
-reference,/data/refdata-cellranger-gex-GRCh38-2020-A
-
-[libraries]
-fastq_id,fastqs,lanes,physical_library_id,feature_types,subsample_rate
-PJB1_GEX,/data/runs/fastqs_gex,any,PJB1,gene expression,
-PJB2_MC,/data/runs/fastqs_mc,any,PJB2,Multiplexing Capture,
-
-[samples]
-sample_id,cmo_ids,description
-PBA,CMO301,PBA
-PBB,CMO302,PBB
-""")
-        config_csv = CellrangerMultiConfigCsv(
-            os.path.join(self.wd,
-                         "10x_multi_config.csv"))
-        self.assertEqual(config_csv.sample_names,['PBA','PBB'])
-        self.assertEqual(config_csv.reference_data_path,
-                         "/data/refdata-cellranger-gex-GRCh38-2020-A")
-        self.assertEqual(config_csv.gex_libraries,
-                         ['PJB1_GEX',])
-        self.assertEqual(config_csv.gex_library('PJB1_GEX'),
-                         { 'fastqs': '/data/runs/fastqs_gex',
-                           'lanes': 'any',
-                           'library_id': 'PJB1',
-                           'feature_type': 'gene expression',
-                           'subsample_rate': ''
-                         })
-        self.assertEqual(config_csv.fastq_dirs,
-                         { 'PJB1_GEX': '/data/runs/fastqs_gex',
-                           'PJB2_MC': '/data/runs/fastqs_mc'
-                         })
-        self.assertEqual(config_csv.pretty_print_samples(),
-                         "PBA, PBB")
-
-    def test_cellranger_multi_config_csv_reduced_fields(self):
-        """
-        CellrangerMultiConfigCsv: check with 'reduced' config.csv
-        """
-        with open(os.path.join(self.wd,"10x_multi_config.csv"),'wt') as fp:
-            fp.write("""[gene-expression]
-reference,/data/refdata-cellranger-gex-GRCh38-2020-A
-
-[libraries]
-fastq_id,fastqs,feature_types
-PJB1_GEX,/data/runs/fastqs_gex,gene expression
-PJB2_MC,/data/runs/fastqs_mc,Multiplexing Capture
-
-[samples]
-sample_id,cmo_ids
-PBA,CMO301
-PBB,CMO302
-""")
-        config_csv = CellrangerMultiConfigCsv(
-            os.path.join(self.wd,
-                         "10x_multi_config.csv"))
-        self.assertEqual(config_csv.sample_names,['PBA','PBB'])
-        self.assertEqual(config_csv.reference_data_path,
-                         "/data/refdata-cellranger-gex-GRCh38-2020-A")
-        self.assertEqual(config_csv.gex_libraries,
-                         ['PJB1_GEX',])
-        self.assertEqual(config_csv.gex_library('PJB1_GEX'),
-                         { 'fastqs': '/data/runs/fastqs_gex',
-                           'lanes': 'any',
-                           'library_id': None,
-                           'feature_type': 'gene expression',
-                           'subsample_rate': ''
-                         })
-        self.assertEqual(config_csv.fastq_dirs,
-                         { 'PJB1_GEX': '/data/runs/fastqs_gex',
-                           'PJB2_MC': '/data/runs/fastqs_mc'
-                         })
-        self.assertEqual(config_csv.pretty_print_samples(),
-                         "PBA, PBB")
